Optionally read fuzz target class from JAR manifest
By reading the fuzz target class from a JAR manifest entry rather than
a command-line argument, fuzz targets can be fully self-contained.

This commit adds a Java function that looks for a unique
`Fuzz-Target-Class` attribute in all manifests on the classpath and
returns it to the driver if found. If no such entry is found, it
falls back to the `--target_class` commandline parameter.